feat: split application into modules
