+++ src/scripts/fillCurrency.js
@@ -1,0 +1,7 @@
+export default function fillCurrency(parent, data) {
+  for (let key in data.rates) {
+    const option = document.createElement('option')
+    option.innerText = key;
+    parent.appendChild(option);
+  }
+}
